simplify: run command and publish response event on single thread instead of having a separate thread pool for publishing
Note: there are now just three core abstractions:
  - the listener which hands off events from a queue to
  - the handler, which generates and then runs command(s) and converts the
    results to response events, which it hands to
  - the publisher to validate and post to queue(s).

Note: I think this doesn't raise any performance concerns, because the
publishing to a queue is nothing compared to running the subprocess jobs.
Tacking on the publishing in the same thread is not likely to hold up the
thread for dealing with new stuff coming in.