pass simple test (test_adapter_mq)
Note: some details of the implementation seem overengineered to me now. Do we
really need the ability for multiple commands to handle a single event? Do
we really need the ability to send responses to different queues depending on
handler logic (thus complicating the clients)?

Note: one key requirement is _the ability to cancel running tasks_ (if they are
still running). For this it would be helpful to have _two responses_ per
event: the first that identifies the task as it begins to execute; and the
second that is the result of the completed task.

These two changes -- cutting what's not needed and adding what is -- guide the
next steps I think.