test reimplementation as ttq
Note: consuming events is now done through a pika connection.sleep loop, a
preferred method to time.sleep and start_consuming running in a separate thread.
Documentation for pika is amazingly opaque and there are bad examples everywhere.
This simplifies things further, so that the listener is not running in a threadm
it's just basically a closure for defining the message callback.

Note also, I was missing some key things like ack'ing messages (!)

Added logging throughout.

Storage of process/correlation_id info still to be done.