remove EventProtocol -> Command from EventProtocol, implement as separate func
  - run command takes `app` parameter passed in from the top

Note: see comments in previous commit for rationale.

Note: this is inescapably app code, not directly representable in config files.
The concept is that both config.events and this function would be derived from
dynamic import of app code in a specified file. The representation would be in
the form of a mapping like this:

    { SomeEvent: SomeCommandFactory(some_config),
      OtherEvent: other_command_function,
    }

which then the CLI interface would derive the events list from and compile down
into a single EventProtocol -> Command function to pass into `run`.