type cleanup using mypy strict mode
Note: the 'EventMapping' interface between app and this library is broken,
in terms of types. It needs some rethinking and simplification. This is
basically what's needed:

    App = Callable[[Context],Command]

where Command is a Protocol or ABC.

For convenience, following a typical convention, your app might define its
commands as a mapping:

    App_ = Mapping[str, Callable[[str,str,bytes],Command]]

where the keys are the `type` names from the Context, and the values are
decode functions from content_type, content_encoding, and message body
to Command.

Or expecting content_type == 'application/json':

    App__ = Mapping[str, Callable[[str,dict[str,Any]],Command]]

ttq should accept the app in the most general form (the first above).
While providing some 'compiler' functions for typical cases like the
second and third.